Create Basic Editor
When I say basic I mean *basic*. You can move your cursor around but it's not fully locked to where text is, newly typed text is always added to the end of the buffer regardless of the location of the cursor. The only vim motions/commands supported are `hjkl` and transitioning modes, and `:q` to exit.

However, I feel I've created a decent base layer to build on top of, and the struct for processing input and handling the current state *should* be somewhat scalable.