Implement Paging
Creating a "window" of visible text via vertical and horizontal offsets. Offset is a position struct utilized by the inputprocessor to track where it should start displaying content,
and when the cursor moves "offscreen" in either axis it is compared to the available text and current offset to see if the offset should be increased or decreased.

This ended up requiring or encouraging some refactors and improvements to how cursor movement is handled.