Add Comments for Concurrent Screen Drawing
Attempted to add wait groups for concurrent screen drawing, it doesn't kill the program but status line is inconsistently drawn.
I think it has to do with the functions finishing in different orders, but I'm not positive and since I'm not seeing performance issues with the current solution it doesn't feel worth the time to keep messing with.

I'll come back to this later, leaving the comments in place as a reminder.